Introduce state container
