+++ src/main.js
@@ -39,6 +39,7 @@ function newWindow(path) {
   })
 
   window.webContents.audioMuted = !setting.get('sound.enable')
+  window.webContents.openDevTools()
 
   window.webContents.on('did-finish-load', () => {
     if (path) window.webContents.send('load-file', path)
