+++ src/modules/dialog.js
@@ -1,9 +1,12 @@
-const {remote} = require('electron')
-const {app, dialog} = remote
-const helper = require('./helper')
-const t = require('../i18n').context('dialog')
+import {remote} from 'electron'
+import i18n from '../i18n.js'
+import sabaki from './sabaki.js'
+import {noop} from './helper.js'
 
-exports.showMessageBox = function(
+const t = i18n.context('dialog')
+const {app, dialog} = remote
+
+export function showMessageBox(
   message,
   type = 'info',
   buttons = [t('OK')],
@@ -24,7 +27,7 @@ exports.showMessageBox = function(
   return result
 }
 
-exports.showFileDialog = function(type, options, callback = helper.noop) {
+export function showFileDialog(type, options, callback = noop) {
   sabaki.setBusy(true)
 
   let [t, ...ype] = [...type]
@@ -39,19 +42,15 @@ exports.showFileDialog = function(type, options, callback = helper.noop) {
   callback({result})
 }
 
-exports.showOpenDialog = function(options, callback) {
-  return exports.showFileDialog('open', options, callback)
+export function showOpenDialog(options, callback) {
+  return showFileDialog('open', options, callback)
 }
 
-exports.showSaveDialog = function(options, callback) {
-  return exports.showFileDialog('save', options, callback)
+export function showSaveDialog(options, callback) {
+  return showFileDialog('save', options, callback)
 }
 
-exports.showInputBox = function(
-  message,
-  onSubmit = helper.noop,
-  onCancel = helper.noop
-) {
+export function showInputBox(message, onSubmit = noop, onCancel = noop) {
   sabaki.setState({
     inputBoxText: message,
     showInputBox: true,
@@ -60,8 +59,8 @@ exports.showInputBox = function(
   })
 }
 
-exports.closeInputBox = function() {
-  let {onInputBoxCancel = helper.noop} = sabaki.state
+export function closeInputBox() {
+  let {onInputBoxCancel = noop} = sabaki.state
   sabaki.setState({showInputBox: false})
   onInputBoxCancel()
 }
